requirements.txt
pyserial
XBee
psutil
influxdb
pyyaml
configparser
binascii
requests
pytz